fpga: launch all syntheses with appropriate number of threads
Rename messages.tcl to common.tcl and add a number of CPU
detection based on the system's getconf with a fallback to "4".
Change all invocations of "launch_run" to use this derived
number of jobs instead of none or a hard-coded number of "8".
To that end, an additional number of .tcl files have to include
common.tcl now.